requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.2.0+cpu
torchvision==0.4.0+cpu
click==7.1.1
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.3
Pillow==6.2.2
Werkzeug==1.0.1